hq_demo/pyproject.toml
[project]
name = "ddnm-hq-demo"
version = "0.1.0"
description = "DDNM High Quality Image Inpainting Demo"
requires-python = ">=3.8"
dependencies = [
    # Core ML framework
    "torch>=2.0.0,<3.0.0",
    "torchvision>=0.15.0",
    # Image processing
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    # Configuration and utilities
    "PyYAML>=6.0",
    "tqdm>=4.64.0",
    "blobfile>=2.0.0",
    "pytorch-fid>=0.3.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
